requirements.txt
langchain>=1.0.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
pydantic>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
apscheduler>=3.10.0
google-api-python-client>=2.150.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
langchain-ollama>=0.1.0